5.10,11- Author page,Attachment and tiny plugin inegrate

+++ post-formats/content.php
@@ -31,42 +31,37 @@
                         </div>
                         <div class="row">
                             <div class="col-md-12">
-                                <!-- <ul class="list-unstyled">
-                                    <li>dhaka</li>
-                                </ul> -->
+                                <div class="slider">
+                                            <?php
+                                            if ( class_exists( 'Attachments' ) ) {
+                                                $attachments = new Attachments( 'slider' );
+                                                if ( $attachments->exist() ) {
+                                                    while ( $attachment = $attachments->get() ) { ?>
+                                                        <div>
+                                                            <?php echo $attachments->image( 'large' ); ?>
+                                                        </div>
+                                                        <?php
+                                                    }
+                                                }
+                                            }
+                                            ?>
+                                        </div>
+                                        <div>
+                                            <?php
+                                            if ( !class_exists( 'Attachments' ) ) {
+                                                if ( has_post_thumbnail() ) {
+                                                    $thumbnail_url = get_the_post_thumbnail_url( null, "large" );
+                                                    printf( '<a class="popup" href="%s" data-featherlight="image">', $thumbnail_url );
+                                                    the_post_thumbnail( "large", array( "class" => "img-fluid" ) );
+                                                    echo '</a>';
+                                                }
+                                            }
 
-                                <!-- showing post tags start -->
-                               <?php echo get_the_tag_list( "<ul class=\"list-unstyled\"><li>", "<li></li>","</li></ul>"); ?>
-                                <!-- showing post tags end -->
-                                 <p>
-                                    <?php
-                                    if(has_post_thumbnail(  )){
-                                        // $thumbnail_image = get_the_post_thumbnail_url(null, "large" );
-                                        echo '<a class="popup" href="#" data-featherlight="image">';
-                                        the_post_thumbnail( "large", array("class"=>"img-fluid") );
-                                        echo "</a>";
-                                    }
-                                    ?>
-                                </p>
-                                <p>
-                                    <!-- checking single page or not start -->
-                                    <?php
+                                            the_content();
+                                            wp_link_pages();
 
-                                       
-                                         the_content(  );
-
-                                         // page break show
-                                         wp_link_pages(  );
-
-                                         next_post_link();
-                                         
-                                         echo "</br>";
-
-                                         previous_post_link();
-                                       
-                                    ?>
-                                    <!-- checking single page or not end -->
-                                </p>
+                                            ?>
+                                        </div>
                             </div>
                             <!-- showing author information start -->
                             <div class="author_information">
